Update REL task.
