+++ spacy_llm/tasks/rel/__init__.py
@@ -1,0 +1,11 @@
+from .examples import RelationItem, RELExample
+from .registry import make_rel_task
+from .task import DEFAULT_REL_TEMPLATE, RELTask
+
+__all__ = [
+    "DEFAULT_REL_TEMPLATE",
+    "make_rel_task",
+    "RelationItem",
+    "RELExample",
+    "RELTask",
+]
