+++ spacy_llm/tasks/summarization/parser.py
@@ -2,7 +2,7 @@ from typing import Iterable
 
 
 def parse_responses_v1(responses: Iterable[str]) -> Iterable[str]:
-    """Parses LLM responses for spacy.Summarization.v1 and maps them onto Docs.
+    """Parses LLM responses for spacy.Summarization.v1.
     responses (Iterable[str]): LLM responses.
     field (str): Field to store responses in.
     RETURNS (Iterable[str]): Summary per doc/response.
